arie-backend/requirements.txt
# ARIE Finance — Python Dependencies
# ═══════════════════════════════════════════════════════════
# Core Framework
tornado>=6.1

# Authentication & Security
bcrypt>=4.0.0
PyJWT>=2.6.0
cryptography>=41.0.0

# Data Validation
pydantic>=2.0.0
typing_extensions>=4.0.0

# HTTP Client (for external API integrations)
requests>=2.28.0

# Environment & Configuration
python-dotenv>=1.0.0

# Production WSGI Server (use with reverse proxy)
gunicorn>=21.2.0

# PostgreSQL Support (production database)
psycopg2-binary>=2.9.0

# AWS S3 Support (document storage)
boto3>=1.28.0

# Observability & Monitoring
prometheus-client>=0.17.0

# Async SQLite Support (for resilience layer)
aiosqlite>=0.17.0

# Server-side PDF Generation (Sprint 3)
weasyprint>=60.0

# ── Testing ───────────────────────────────────────────────
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
